requirements.txt
streamlit
google-genai
pandas 
python-dotenv
psycopg2-binary